% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{adjalp}
\alias{adjalp}
\title{obtain the residual statistics}
\usage{
adjalp(
  Y.sim,
  FAP0 = 0.3,
  side = "two-sided",
  interval = c(1e-05, 0.4),
  tol = 1e-06
)
}
\arguments{
\item{Y.sim}{is the transformed vector}

\item{FAP0}{is the matrix of laggy coefficients}

\item{interval}{is the interval}

\item{tol}{is the tolerance}
}
\description{
obtain the residual statistics
}
