% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{adjalpha.ph1}
\alias{adjalpha.ph1}
\title{obtain the residual statistics}
\usage{
adjalpha.ph1(
  Y.hat,
  sigma2.hat,
  Y.sim,
  FAP0 = 0.3,
  side = "two-sided",
  tol = 1e-06
)
}
\arguments{
\item{Y.hat}{is the transformed vector}

\item{sigma2.hat}{is the transformed vector}

\item{Y.sim}{is the transformed vector}

\item{FAP0}{is the matrix of laggy coefficients}

\item{side}{is the side}

\item{tol}{is the tolerance}
}
\description{
obtain the residual statistics
}
