% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{adjalpha.ph2}
\alias{adjalpha.ph2}
\title{obtain the residual statistics}
\usage{
adjalpha.ph2(
  Y.hat,
  sigma2.hat,
  Y.sim,
  cs.mean,
  cs.sd,
  ARL0 = 360,
  side = "two-sided",
  tol = 1e-06
)
}
\arguments{
\item{Y.hat}{is the transformed vector}

\item{sigma2.hat}{is the transformed vector}

\item{Y.sim}{is the transformed vector}

\item{ARL0}{is the matrix of laggy coefficients}

\item{side}{is the side}

\item{tol}{is the tolerance}
}
\description{
obtain the residual statistics
}
