% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{backtrans}
\alias{backtrans}
\title{Back-transform the data}
\usage{
backtrans(Y, log = TRUE, const = 1, sta = TRUE, meanY = 0, sdY = 1)
}
\arguments{
\item{Y}{is the input}

\item{log}{is the flag triggering the log transformation}

\item{const}{is the constant added to the input during the log transformation}

\item{sta}{is the flag triggering the standardization}

\item{meanY}{is the mean of the original Y}

\item{sdY}{is the standard deviation of the original Y}
}
\description{
gets the back-transformed input
}
\examples{
alpha <- c(0.03083069, 0.06242601, 0.09120189)
lambda <- 0.239385
pi <- 0.1453097

TT <- 183
w <- 28
Y <- rzinpoisinar3(TT + w, alpha, lambda, pi, ceiling(TT / 2) + w, delta = 1, burnin = 100)
ma <- movaver(Y, w)
ma.tr <- trans(ma, TRUE, 1, TRUE)
ma.batr <- backtrans(ma.tr$Y, TRUE, 1, TRUE, ma.tr$meanY, ma.tr$sdY)
}
