% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring-package.R
\docType{package}
\name{BayesianLASSOMonitoring-package}
\alias{BayesianLASSOMonitoring-package}
\alias{_PACKAGE}
\title{BayesianLASSOMonitoring: Bayesian LASSO Random Flexible Level Shift Model Based Phase I Monitoring}
\description{
A Phase I Control Chart which is built using Random Flexible Level Shift Model is provided. The model is trained using the Bayesian LASSO Technique.

Package description.
}
\references{
McCulloch, R. E., & Tsay, R. S. (1993). Bayesian inference and prediction for mean and variance shifts in autoregressive time series. Journal of the american Statistical association, 88(423), 968-978.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bolus123/BayesianLASSOMonitoring}
  \item Report bugs at \url{https://github.com/bolus123/BayesianLASSOMonitoring/issues}
}

}
\keyword{package}
