% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOPh1.R
\name{BayesianLASSOPh1}
\alias{BayesianLASSOPh1}
\title{Bayesian LASSO Phase I Monitoring}
\usage{
BayesianLASSOPh1(
  Y,
  H = NULL,
  X = NULL,
  q = 5,
  A = diag(nrow = q + ifelse(is.null(X), 0, dim(X)[2])),
  a = 0.1,
  b = 0.1,
  alpha = 0.1,
  beta = 0.1,
  theta1 = 1,
  theta2 = 1,
  xi2 = 0.1,
  method = "MonoALASSO",
  bound0 = Inf,
  boundqplus1 = 0,
  nsim = 1000,
  by = 1,
  burnin = 1000,
  tol = 1e-10,
  standardized = TRUE,
  FAP0 = 0.2,
  estimation.PPP = "median",
  nsim.PPP = 1000
)
}
\arguments{
\item{Y}{is a vector.}

\item{H}{is the design matrix for shifts.}

\item{X}{is the input matrix}

\item{q}{is the number of lags.}

\item{A}{is a given variance-covariance matrix in MT and regression for the slab-and-spike coefficients.}

\item{a}{is a given shape of the prior gamma distribution for sigma2.}

\item{b}{is a given scale of the prior gamma distribution for sigma2.}

\item{alpha}{is a given shape of the prior gamma distribution for lambda2.}

\item{beta}{is a given scale of the prior gamma distribution for lambda2.}

\item{theta1}{is a given shape1 of the prior beta distribution for the probability of Tau and Kappa.}

\item{theta2}{is a given shape2 of the prior beta distribution for the probability of Tau and Kappa.}

\item{xi2}{is a given variance of the prior normal distribution for shifts.}

\item{method}{is a choice of methods including MT(McCulloch-Tsay), regression, LASSO, ALASSO(Adaptive LASSO), MonoLASSO(LASSO with Monotonicity constrains), MonoALASSO(Adaptive LASSO with Monotonicity constrains).}

\item{bound0}{is an upper bound of the methods with Monotonicity constrains.}

\item{boundqplus1}{is  a lower bound of the methods with Monotonicity constrains.}

\item{nsim}{is the number of draws from MCMC.}

\item{by}{is the interval of systematic sampling for the draws from MCMC.}

\item{burnin}{is the length of burn-in period.}

\item{tol}{is the tolerance level.}

\item{standardized}{is the flag triggering the standardization for the time series}

\item{FAP0}{is the given false alarm probability}

\item{estimation.PPP}{is the estimation for Mu0, Phi and sigma2}

\item{nsim.PPP}{is the number of draws for PPP}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- BayesianLASSOPh1(Y, H = H, q = q, nsim = nsim, burnin = burnin)

}
