\name{BayesianMonitoring1-package}
\alias{BayesianMonitoring1-package}
\alias{BayesianMonitoring1}
\docType{package}
\title{
\packageTitle{BayesianMonitoring1}
}
\description{
\packageDescription{BayesianMonitoring1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesianMonitoring1}
\packageIndices{BayesianMonitoring1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BayesianMonitoring1}

Maintainer: \packageMaintainer{BayesianMonitoring1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
