\name{BenjaminiHochberg}
\alias{BenjaminiHochberg}
\title{BenjaminiHochberg}
\description{
  BenjaminiHochberg
}
\usage{
BenjaminiHochberg(FDR, beta2, side, KernelSmoothing, bandwidth)
}
\arguments{
  \item{FDR}{a numeric vector}
  \item{beta2}{a numeric vector}
  \item{side}{a numeric vector}
  \item{KernelSmoothing}{a numeric vector}
  \item{bandwidth}{a numeric vector}
}
\value{
  value
  
}
\details{
  details
}
\references{
  no reference
}
\examples{
  \dontrun{}
}
\author{Yuhui Yao}

