% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHProcRFLSM.R
\name{BHProcRFLSM}
\alias{BHProcRFLSM}
\title{do the multiple testing for all shifts using Benjamini-Hochberg procedure}
\usage{
BHProcRFLSM(TauGamma, FDR0 = 0.1, tail = "2-sided", method = "raw")
}
\arguments{
\item{TauGamma}{is the distributions of shifts.}

\item{FDR0}{is a given constant for FDR.}

\item{tail}{is type of the test.}

\item{method}{do the procedure with or without kernel smoothing.}
}
\description{
do the multiple testing for all shifts using Benjamini-Hochberg procedure
}
\examples{
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, q, diag(nrow = q), 0.1, 0.1, 0.1, 0.1, 
1, 1, 0.1, "MonoALASSO", Inf, 0, 1000, 1, 100, 1e-10, H)

BHProcRFLSM(result$Tau * result$Gamma)
}
