\name{BonferroniCorrection}
\alias{BonferroniCorrection}
\title{BonferroniCorrection}
\description{
  BonferroniCorrection
}
\usage{
BonferroniCorrection(FAP, beta2, side)
}
\arguments{
  \item{FAP}{a numeric vector}
  \item{beta2}{a numeric vector}
  \item{side}{a numeric vector}
}
\value{
  value
  
}
\details{
  details
}
\references{
  no reference
}
\examples{
  \dontrun{}
}
\author{Yuhui Yao}

