% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{CompResamp1}
\alias{CompResamp1}
\title{comparison using resampling}
\usage{
CompResamp1(
  Y,
  Phi,
  muq,
  sigma2,
  FAP0 = 0.2,
  interval = c(0.51, 1),
  nsim = 1e+05,
  tol = 1e-06
)
}
\arguments{
\item{FAP0}{is FAP0.}

\item{P}{is P}

\item{methodP}{is methodP}

\item{methodComp}{is methodComp.}
}
\description{
comparison using resampling
}
\examples{
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, q, diag(nrow = q), 0.1, 0.1, 0.1, 0.1, 
1, 1, 0.1, "MonoALASSO", Inf, 0, 1000, 1, 100, 1e-10, H)

MultiComp(result$p)

}
