% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinomial.R
\name{dtrinomial}
\alias{dtrinomial}
\title{Get a pmf of trinomial distribution}
\usage{
dtrinomial(x, n, p0)
}
\arguments{
\item{x}{is x}

\item{n}{is trials}

\item{p0}{is ties}
}
\description{
gets the simulated data
}
