% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{Fit0}
\alias{Fit0}
\title{obtain the fits under H0}
\usage{
Fit0(Y, Phi, muq)
}
\arguments{
\item{Y}{is a vector}

\item{Phi}{is the coefficient}

\item{muq}{is the mean}
}
\description{
obtain the fits under H0
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

Fit0(Y, result$Phi, result$muq)

}
