% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFDR.R
\name{getH0}
\alias{getH0}
\title{get a true proportion of H0}
\usage{
getH0(
  TauGamma,
  tail = "2-sided",
  FDR.method = "meinshausen",
  pvalue.method = "raw"
)
}
\arguments{
\item{TauGamma}{is the distributions of shifts.}

\item{tail}{is type of the test.}

\item{method}{get p values with or without kernel smoothing.}
}
\description{
get a true proportion of H0
}
\examples{
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, q, diag(nrow = q), 0.1, 0.1, 0.1, 0.1, 
1, 1, 0.1, "MonoALASSO", Inf, 0, 1000, 1, 100, 1e-10, H)

getH0(result$Tau * result$Gamma)
}
