% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHMatCycle}
\alias{getHMatCycle}
\title{get a design matrix as that in MT}
\usage{
getHMatCycle(T, q, wt)
}
\arguments{
\item{T}{is length of a process.}

\item{q}{is the number of lags.}

\item{wt}{is the waiting time.}
}
\description{
get a design matrix as that in MT
}
\examples{
getHMatMT(100, 5)
}
