% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHMatIsolated}
\alias{getHMatIsolated}
\title{get a design matrix for isolated shift}
\usage{
getHMatIsolated(T, q, w)
}
\arguments{
\item{T}{is length of a process.}

\item{q}{is the number of lags.}

\item{w}{is the subgroup size.}
}
\description{
get a design matrix for isolated shift
}
\examples{
getHMatIsolated(100, 5, 1)
}
