% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHMatMT}
\alias{getHMatMT}
\title{Design Matrix (MT)}
\usage{
getHMatMT(T, q)
}
\arguments{
\item{T}{is length of a process.}

\item{q}{is the number of lags.}
}
\description{
gets a design matrix as that in MT
}
\examples{
getHMatMT(100, 5)
}
