% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHMatSustained}
\alias{getHMatSustained}
\title{get a design matrix for sustained shift}
\usage{
getHMatSustained(T, q, w)
}
\arguments{
\item{T}{is length of a process.}

\item{q}{is the number of lags.}

\item{w}{is the subgroup size.}
}
\description{
get a design matrix for sustained shift
}
\examples{
getHMatSustained(100, 5, 1)
}
