\name{getPosterior}
\alias{getPosterior}
\title{getPosterior}
\description{
  getPosterior
}
\usage{
getPosterior(Y, V, X, lambda2, beta0, beta1, beta2, burnin, nsim)
}
\arguments{
  \item{Y}{a numeric vector}
  \item{V}{a numeric vector}
  \item{X}{a numeric vector}
  \item{lambda2}{a numeric vector}
  \item{beta0}{a numeric vector}
  \item{beta1}{a numeric vector}
  \item{beta2}{a numeric vector}
  \item{burnin}{a numeric vector}
  \item{nsim}{a numeric vector}
}
\value{
  value
  
}
\details{
  details
}
\references{
  no reference
}
\examples{
  \dontrun{}
}
\author{Yuhui Yao}

