\name{getPosteriorH0}
\alias{getPosteriorH0}
\title{getPosteriorH0}
\description{
  getPosteriorH0
}
\usage{
getPosteriorH0(Y, V, lambda2, beta0, beta1, burnin, nsim)
}
\arguments{
  \item{Y}{a numeric vector}
  \item{V}{a numeric vector}
  \item{lambda2}{a numeric vector}
  \item{beta0}{a numeric vector}
  \item{beta1}{a numeric vector}
  \item{burnin}{a numeric vector}
  \item{nsim}{a numeric vector}
}
\value{
  value
  
}
\details{
  details
}
\references{
  no reference
}
\examples{
  \dontrun{}
}
\author{Yuhui Yao}

