% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{getPvalueKSRFLSM}
\alias{getPvalueKSRFLSM}
\title{get the P value for RFLSM with kernel smoothing}
\usage{
getPvalueKSRFLSM(TauGamma, tail = "2-sided")
}
\arguments{
\item{TauGamma}{is the distributions of shifts.}

\item{tail}{is type of the test.}
}
\description{
get the P value for RFLSM with kernel smoothing
}
