% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getV}
\alias{getV}
\title{get matrix V}
\usage{
getV(Y, q)
}
\arguments{
\item{Y}{is a vector.}

\item{q}{is a number of lags.}
}
\description{
get matrix V
}
\examples{
Y <- arima.sim(list(ar = 0.5), 100)
getV(Y, 10)
}
