% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getXSeasonalityFS}
\alias{getXSeasonalityFS}
\title{gets a design matrix for Fourier series}
\usage{
getXSeasonalityFS(T, s, n)
}
\arguments{
\item{T}{is length of a process.}

\item{s}{is the number of period}

\item{n}{is the number of Fourier series}
}
\description{
gets a design matrix for Fourier series
}
\examples{
getXSeasonalityFS(100, 15, 10)
}
