% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsRFLSM.R
\name{GibbsRFLSM.ma}
\alias{GibbsRFLSM.ma}
\title{Random Flexible Level Shift Model}
\usage{
GibbsRFLSM.ma(
  Y,
  w = 7,
  H = NULL,
  X = NULL,
  Y0 = rep(mean(Y), w - 1),
  q = 5,
  A = diag(nrow = q),
  a = 0.1,
  b = 0.1,
  alpha = 0.1,
  beta = 0.1,
  theta1 = 1,
  theta2 = 1,
  xi2 = 0.1,
  method = "MonoALASSO",
  bound0 = Inf,
  boundqplus1 = 0,
  nsim = 1000,
  by = 1,
  burnin = 1000,
  tol = 1e-10,
  log = TRUE,
  const = 1,
  sta = TRUE
)
}
\arguments{
\item{Y}{is a vector.}

\item{H}{is the design matrix for shifts.}

\item{X}{is the input matrix}

\item{q}{is the number of lags.}

\item{A}{is a given variance-covariance matrix in MT and regression for the slab-and-spike coefficients.}

\item{a}{is a given shape of the prior gamma distribution for sigma2.}

\item{b}{is a given scale of the prior gamma distribution for sigma2.}

\item{alpha}{is a given shape of the prior gamma distribution for lambda2.}

\item{beta}{is a given scale of the prior gamma distribution for lambda2.}

\item{theta1}{is a given shape1 of the prior beta distribution for the probability of Tau and Kappa.}

\item{theta2}{is a given shape2 of the prior beta distribution for the probability of Tau and Kappa.}

\item{xi2}{is a given variance of the prior normal distribution for shifts.}

\item{method}{is a choice of methods including MT(McCulloch-Tsay), regression, LASSO, ALASSO(Adaptive LASSO), MonoLASSO(LASSO with Monotonicity constrains), MonoALASSO(Adaptive LASSO with Monotonicity constrains).}

\item{bound0}{is an upper bound of the methods with Monotonicity constrains.}

\item{boundqplus1}{is  a lower bound of the methods with Monotonicity constrains.}

\item{nsim}{is the number of draws from MCMC.}

\item{by}{is the interval of systematic sampling for the draws from MCMC.}

\item{burnin}{is the length of burn-in period.}

\item{tol}{is the tolerance level.}

\item{log}{is the flag triggering the log transformation}

\item{const}{is the constant added to the input during the log transformation}

\item{sta}{is the flag triggering the standardization}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
\examples{
alpha <- c(0.03083069, 0.06242601, 0.09120189)
lambda <- 0.239385
pi <- 0.1453097

TT <- 183
w <- 28
Y <- rzinpoisinar3(TT + w, alpha, lambda, pi, ceiling(TT / 2) + w, delta = 1, burnin = 100)

H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM.ma(Y, H = H, q = q, nsim = nsim, burnin = burnin)

}
\references{
McCulloch, R. E., & Tsay, R. S. (1993). Bayesian inference and prediction for mean and variance shifts in autoregressive time series. Journal of the american Statistical association, 88(423), 968-978.
}
