% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{GibbsRFLSM.max.residual.sim}
\alias{GibbsRFLSM.max.residual.sim}
\title{simulates the maximums of residuals}
\usage{
GibbsRFLSM.max.residual.sim(
  Y,
  Phi,
  muq,
  sigma2,
  Phihat,
  muqhat,
  sigma2hat,
  logcc = FALSE,
  standardization = FALSE,
  nsim = 1000
)
}
\arguments{
\item{Y}{is a vector}

\item{Phi}{is the coefficient}

\item{sigma2}{is the variance of errors}

\item{Phihat}{is the coefficient}

\item{sigma2hat}{is the variance of errors}

\item{nsim}{is the number of simulations}

\item{Mu}{is the mean}

\item{Muhat}{is the mean}
}
\description{
simulates the maximums of residuals
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

Phihat <- rep(NA, q)
for (i in 1:q) {
  Phihat[i] <- median(result$Phi[i, ])
}

Muhat <- rep(NA, T)
for (i in 1:T) {
  Muhat[i] <- median(result$Mu[i, ])
}

sigma2hat <- median(result$sigma2)

GibbsRFLSM.simmax.residual(Y, result$Phi, result$Mu, result$sigma2, 
Phihat, Muhat, sigma2hat)

}
