% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{GibbsRFLSM.max.scan}
\alias{GibbsRFLSM.max.scan}
\title{simulates the maximums of retrospective scan statistics}
\usage{
GibbsRFLSM.max.scan(
  Y,
  w,
  Phihat,
  muqhat,
  logcc = FALSE,
  standardization = FALSE,
  tol = 1e-10
)
}
\arguments{
\item{Y}{is a vector}

\item{w}{is the moving window}

\item{Phihat}{is the coefficient}

\item{N}{is the sum of all counts}

\item{Phi}{is the coefficient}

\item{Mu}{is the mean}

\item{sigma2}{is the variance of errors}

\item{Muhat}{is the mean}

\item{nsim}{is the number of simulations}
}
\description{
simulates the maximums of retrospective scan statistics
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

Phihat <- rep(NA, q)
for (i in 1:q) {
  Phihat[i] <- median(result$Phi[i, ])
}

Muhat <- rep(NA, T)
for (i in 1:T) {
  Muhat[i] <- median(result$Mu[i, ])
}

sigma2hat <- median(result$sigma2)

GibbsRFLSM.simmax.residual(Y, result$Phi, result$Mu, result$sigma2, 
Phihat, Muhat)

}
