% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{GibbsRFLSM.PPP.residual}
\alias{GibbsRFLSM.PPP.residual}
\title{simulates posterior predictive p values using for the Phase I chart}
\usage{
GibbsRFLSM.PPP.residual(
  Y,
  Phi,
  Mu0,
  sigma2,
  Phihat,
  Mu0hat,
  sigma2hat,
  nsim = 1000
)
}
\arguments{
\item{Y}{is a vector}

\item{Phi}{is the coefficient}

\item{Mu0}{is the mean}

\item{sigma2}{is the variance of errors}

\item{Phihat}{is the coefficient}

\item{Mu0hat}{is the mean}

\item{sigma2hat}{is the variance of errors}

\item{nsim}{is the number of simulations}
}
\description{
simulates posterior predictive p values using for the Phase I chart
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

Mu0 <- matrix(NA, nrow = T, ncol = nsim)
for (j in 1:nsim) {
  Mu0[, j] <- result$muq[j]
}

Mu0hat <- rep(NA, T)
for (i in 1:T) {
  Mu0hat[i] <- median(Mu0[i, ])
}

Phihat <- rep(NA, q)
for (i in 1:q) {
  Phihat[i] <- median(result$Phi[i, ])
}

sigma2hat <- median(result$sigma2)

GibbsRFLSM.PPP.residual(Y, result$Phi, Mu0, result$sigma2, 
Phihat, Mu0hat, sigma2hat)

}
