% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsRFLSM}
\alias{GibbsRFLSM}
\title{get a posterior sample using gibbs sampling for Random Flexible Level Shift Model}
\usage{
GibbsRFLSM(
  Y,
  q,
  A,
  a,
  b,
  alpha,
  beta,
  theta1,
  theta2,
  xi2,
  method,
  bound0,
  boundqplus1,
  nsim,
  by,
  burnin,
  tol,
  H = NULL
)
}
\arguments{
\item{Y}{is a vector.}

\item{q}{is the number of lags.}

\item{A}{is a given variance-covariance matrix in MT and regression for the coefficients for autoregressors.}

\item{a}{is a given shape of the prior gamma distribution for sigma2.}

\item{b}{is a given scale of the prior gamma distribution for sigma2.}

\item{alpha}{is a given shape of the prior gamma distribution for lambda2.}

\item{beta}{is a given scale of the prior gamma distribution for lambda22.}

\item{theta1}{is a given shape1 of the prior beta distribution for the probability of Tau.}

\item{theta2}{is a given shape2 of the prior beta distribution for the probability of Tau.}

\item{xi2}{is a given variance of the prior normal distribution for shifts.}

\item{method}{is a choice of methods including MT(McCulloch-Tsay), regression, LASSO, ALASSO(Adaptive LASSO), MonoLASSO(LASSO with Monotonicity constrains), MonoALASSO(Adaptive LASSO with Monotonicity constrains).}

\item{bound0}{is a lower bound of the methods with Monotonicity constrains.}

\item{boundqplus1}{is  a upper bound of the methods with Monotonicity constrains.}

\item{nsim}{is the number of samples draw from MCMC.}

\item{by}{is the interval of systematic sampling for the draws from MCMC.}

\item{burnin}{is the length of burn-in period.}

\item{tol}{is the tolerance.}

\item{H}{is the design matrix for shifts.}
}
\description{
get a posterior sample using gibbs sampling for Random Flexible Level Shift Model
}
\examples{
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, q, diag(nrow = q), 0.1, 0.1, 0.1, 0.1, 1, 1, 0.1, "MonoALASSO", 0, Inf, 1000, 1, 100, 1e-10, H)

}
