% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsRFLSM.sim.R
\name{GibbsRFLSM.sim.ph1.H0}
\alias{GibbsRFLSM.sim.ph1.H0}
\title{Get a simulation transformed data using Random Flexible Level Shift Model in the retrospective phase under H0 without censoring}
\usage{
GibbsRFLSM.sim.ph1.H0(
  nsim,
  n,
  Phi,
  muq,
  sigma2,
  X = NULL,
  Beta = NULL,
  Kappa = NULL,
  Y1 = rep(median(muq), n)
)
}
\arguments{
\item{nsim}{is the number of simulations}

\item{n}{is the length of simulated process}

\item{Phi}{is the matrix of laggy coefficients}

\item{muq}{is the vector of grand mean}

\item{sigma2}{is the vector of variance}

\item{X}{is the matrix of X}

\item{Beta}{is the matrix of coefficients}

\item{Kappa}{is the matrix of triggering the corresponding beta}

\item{Y1}{is the transformed vector}
}
\description{
gets the simulated data
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

}
