% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsRFLSM.sim.R
\name{GibbsRFLSM.sim.ph2}
\alias{GibbsRFLSM.sim.ph2}
\title{Get a simulation transformed data using Random Flexible Level Shift Model in the retrospective phase under H0}
\usage{
GibbsRFLSM.sim.ph2(
  h,
  nsim,
  Phi,
  muq,
  sigma2,
  X = NULL,
  Beta = NULL,
  Kappa = NULL,
  H = NULL,
  Gamma = NULL,
  Tau = NULL,
  Y1 = rep(0, dim(Phi)[1]),
  X1 = NULL,
  H1 = NULL
)
}
\arguments{
\item{nsim}{is the number of simulations}

\item{Phi}{is the matrix of laggy coefficients}

\item{muq}{is the vector of grand mean}

\item{sigma2}{is the vector of variance}

\item{X}{is the matrix of X}

\item{Beta}{is the matrix of coefficients}

\item{Kappa}{is the matrix of triggering the corresponding beta}

\item{H}{is the matrix of X}

\item{Gamma}{is the matrix of coefficients}

\item{Tau}{is the matrix of triggering the corresponding beta}

\item{Y1}{is the transformed vector}
}
\description{
gets the simulated data
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

}
