% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{GibbsRFLSM.sim}
\alias{GibbsRFLSM.sim}
\title{simulates the time series using Draws from MCMC}
\usage{
GibbsRFLSM.sim(Y, Phihat, Muhat, sigma2hat)
}
\arguments{
\item{Y}{is a vector}

\item{Phihat}{is the coefficient}

\item{Muhat}{is the mean}

\item{sigma2hat}{is the variance of errors}
}
\description{
simulates the time series using Draws from MCMC
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

Phihat <- result$Phi[, 1]
Muhat <- result$Mu[, 1]
sigma2hat <- result$sigma2[1]

GibbsRFLSM.sim(Y, Phihat, Muhat, sigma2hat) 

}
