% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{GibbsRFLSM.sim0.ph2}
\alias{GibbsRFLSM.sim0.ph2}
\title{simulates the time series using Draws from MCMC in Phase II under H0}
\usage{
GibbsRFLSM.sim0.ph2(
  n,
  Phi,
  muq,
  sigma2,
  Y0,
  logcc = FALSE,
  standardization = FALSE,
  nsim = 1000
)
}
\arguments{
\item{n}{is the number of time series}

\item{Phi}{is the coefficient}

\item{muq}{is the mean}

\item{sigma2}{is the variance of errors}

\item{logcc}{is the log}

\item{standardization}{is the standardization}

\item{nsim}{is the nsim}

\item{Y}{is a vector}
}
\description{
simulates the time series using Draws from MCMC in Phase II under H0
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

GibbsRFLSM.sim0.ph2(Y, result$Phi, result$muq, result$sigma2) 

}
