% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsRFLSM.R
\name{GibbsRFLSMX}
\alias{GibbsRFLSMX}
\title{Random Flexible Level Shift Model}
\usage{
GibbsRFLSMX(
  Y,
  X = NULL,
  H = NULL,
  bset = bset(betap = ifelse(class(X)[1] == "matrix", diag(nrow = dim(X)[2]), diag(nrow =
    1))),
  tol = 1e-10,
  nsim = 300,
  thin = 10,
  burnin = 1000,
  verbose = TRUE
)
}
\arguments{
\item{Y}{is a vector.}

\item{X}{is the input matrix}

\item{H}{is the design matrix for shifts.}

\item{tol}{is the tolerance level.}

\item{nsim}{is the number of draws from MCMC.}

\item{burnin}{is the length of burn-in period.}

\item{q}{is the number of lags.}

\item{A}{is a given variance-covariance matrix in MT and regression for the slab-and-spike coefficients.}

\item{a}{is a given shape of the prior gamma distribution for sigma2.}

\item{b}{is a given scale of the prior gamma distribution for sigma2.}

\item{alpha}{is a given shape of the prior gamma distribution for lambda2.}

\item{beta}{is a given scale of the prior gamma distribution for lambda2.}

\item{theta1}{is a given shape1 of the prior beta distribution for the probability of Tau and Kappa.}

\item{theta2}{is a given shape2 of the prior beta distribution for the probability of Tau and Kappa.}

\item{xi2}{is a given variance of the prior normal distribution for shifts.}

\item{method}{is a choice of methods including MT(McCulloch-Tsay), regression, LASSO, ALASSO(Adaptive LASSO), MonoLASSO(LASSO with Monotonicity constrains), MonoALASSO(Adaptive LASSO with Monotonicity constrains).}

\item{bound0}{is an upper bound of the methods with Monotonicity constrains.}

\item{boundqplus1}{is  a lower bound of the methods with Monotonicity constrains.}

\item{by}{is the interval of systematic sampling for the draws from MCMC.}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

}
\references{
McCulloch, R. E., & Tsay, R. S. (1993). Bayesian inference and prediction for mean and variance shifts in autoregressive time series. Journal of the american Statistical association, 88(423), 968-978.
}
