% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsRFLSMXcpp}
\alias{GibbsRFLSMXcpp}
\title{Absolute-value-constrained normal distribution}
\usage{
GibbsRFLSMXcpp(
  Y,
  bset,
  tol,
  nsim,
  thin,
  burnin,
  verbose,
  X = NULL,
  H = NULL,
  lambda2 = NULL,
  theta = NULL
)
}
\arguments{
\item{n}{is sample size.}
}
\description{
gets a sample from a normal distribution whose absolute observations are constrained.
}
\examples{
rtwosegnorm(10, 1, 2, 0, 1)
}
