% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{lim.ph1}
\alias{lim.ph1}
\title{obtain the residual statistics}
\usage{
lim.ph1(Y.sim, model, FAP0 = 0.3, side = "two-sided")
}
\arguments{
\item{Y.sim}{is the transformed vector}

\item{FAP0}{is the matrix of laggy coefficients}

\item{side}{is the side}

\item{Y.hat}{is the transformed vector}

\item{sigma2.hat}{is the transformed vector}

\item{tol}{is the tolerance}
}
\description{
obtain the residual statistics
}
