% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{ll}
\alias{ll}
\title{obtain the log likelihood}
\usage{
ll(Y, Phi, Mu, sigma2)
}
\arguments{
\item{Y}{is a vector}

\item{Phi}{is the coefficient}

\item{Mu}{is the mean}

\item{sigma2}{is the variance}
}
\description{
obtain the log likelihood
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

ll(Y, result$Phi, result$Mu, result$sigma2)

}
