% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{movaver}
\alias{movaver}
\title{Caculate the moving averages}
\usage{
movaver(Y, w = 5)
}
\arguments{
\item{Y}{is the input}

\item{w}{is the length of moving window}
}
\description{
gets the moving averages
}
\examples{
alpha <- c(0.03083069, 0.06242601, 0.09120189)
lambda <- 0.239385
pi <- 0.1453097

TT <- 183
w <- 28
Y <- rzinpoisinar3(TT + w, alpha, lambda, pi, ceiling(TT / 2) + w, delta = 1, burnin = 100)
ma <- movaver(Y, w)
}
