% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{MultiComp}
\alias{MultiComp}
\title{check significance using Sidak Correction}
\usage{
MultiComp(P, FAP0 = 0.2, methodP = "median", methodComp = "Sidak")
}
\arguments{
\item{P}{is P}

\item{FAP0}{is FAP0.}

\item{methodP}{is methodP}

\item{methodComp}{is methodComp.}
}
\description{
check significance using Sidak Correction
}
\examples{
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, q, diag(nrow = q), 0.1, 0.1, 0.1, 0.1, 
1, 1, 0.1, "MonoALASSO", Inf, 0, 1000, 1, 100, 1e-10, H)

MultiComp(result$p)

}
