% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring.R
\name{Ph1MultipleTesting.BF}
\alias{Ph1MultipleTesting.BF}
\title{Bayesian LASSO Phase I Monitoring}
\usage{
Ph1MultipleTesting.BF(model, nsim = 1000, FAP0 = 0.2)
}
\arguments{
\item{model}{is model.}

\item{sign.method}{is .}

\item{adj.method}{is}

\item{side}{is side}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
