% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring.R
\name{Ph1MultipleTesting.GammaZNBC}
\alias{Ph1MultipleTesting.GammaZNBC}
\title{Bayesian LASSO Phase I Monitoring}
\usage{
Ph1MultipleTesting.GammaZNBC(
  model,
  w = 7,
  FAP0 = 0.05,
  method = "holm",
  side = "right-sided"
)
}
\arguments{
\item{model}{is model.}

\item{FAP0}{is}

\item{nsim}{is .}

\item{log}{is model.}

\item{const}{is .}

\item{sta}{is}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
