% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring.R
\name{Ph1MultipleTesting}
\alias{Ph1MultipleTesting}
\title{Bayesian LASSO Phase I Monitoring}
\usage{
Ph1MultipleTesting(
  model,
  sign.method = "DM",
  adj.method = "holm",
  side = "two-sided"
)
}
\arguments{
\item{model}{is model.}

\item{sign.method}{is .}

\item{adj.method}{is}

\item{side}{is side}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
