% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring.R
\name{Ph1MultipleTesting.resi}
\alias{Ph1MultipleTesting.resi}
\title{Bayesian LASSO Phase I Monitoring}
\usage{
Ph1MultipleTesting.resi(
  model,
  interval = c(1e-08, 0.5 - 1e-08),
  log = FALSE,
  const = 1,
  sta = FALSE,
  meanY = NULL,
  sdY = NULL,
  nsim = 10000,
  FAP0 = 0.2,
  side = "right-sided",
  tol = 1e-08
)
}
\arguments{
\item{model}{is model.}

\item{log}{is model.}

\item{const}{is .}

\item{sta}{is}

\item{nsim}{is .}

\item{FAP0}{is}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
