% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring.R
\name{Ph1MultipleTesting.Y0}
\alias{Ph1MultipleTesting.Y0}
\title{Bayesian LASSO Phase I Monitoring}
\usage{
Ph1MultipleTesting.Y0(
  model,
  nsim = 1000,
  FAP0 = 0.2,
  log = FALSE,
  const = 1,
  sta = FALSE,
  lowerbound = 0,
  side = "right-sided"
)
}
\arguments{
\item{model}{is model.}

\item{nsim}{is .}

\item{FAP0}{is}

\item{log}{is model.}

\item{const}{is .}

\item{sta}{is}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
