% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring.R
\name{Ph1MultipleTesting.Y01L1CUMSUM}
\alias{Ph1MultipleTesting.Y01L1CUMSUM}
\title{Bayesian LASSO Phase I Monitoring}
\usage{
Ph1MultipleTesting.Y01L1CUMSUM(
  model,
  w = 7,
  FAP0 = 0.2,
  side = "right-sided",
  updateYJ = 1,
  leftcensoring = 1,
  rounding = 1,
  eps = 1e-32,
  backtr = 1,
  nsim = 10000,
  interval = c(1e-06, 0.499999),
  verbose = 0
)
}
\arguments{
\item{model}{is model.}

\item{FAP0}{is}

\item{nsim}{is .}

\item{log}{is model.}

\item{const}{is .}

\item{sta}{is}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
