% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring.R
\name{Ph1MultipleTesting.Y0tr}
\alias{Ph1MultipleTesting.Y0tr}
\title{Bayesian LASSO Phase I Monitoring}
\usage{
Ph1MultipleTesting.Y0tr(
  model,
  FAP0 = 0.2,
  side = "right-sided",
  nsim = 10000,
  interval = c(1e-06, 0.499999)
)
}
\arguments{
\item{model}{is model.}

\item{FAP0}{is}

\item{nsim}{is .}

\item{log}{is model.}

\item{const}{is .}

\item{sta}{is}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
