% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring.R
\name{Ph2BayesianLASSO.EWMA}
\alias{Ph2BayesianLASSO.EWMA}
\title{Bayesian LASSO Phase I Monitoring}
\usage{
Ph2BayesianLASSO.EWMA(
  Y,
  Ph1BayesianLASSO.model,
  lambda = 0.05,
  w = 28,
  H = NULL,
  X = NULL,
  Y1 = rep(0, dim(Ph1BayesianLASSO.model$Phi)[1] + w),
  X1 = NULL,
  H1 = NULL,
  log = TRUE,
  const = 1,
  sta = TRUE,
  meanY = 0,
  sdY = 1,
  Y.hat.method = "median",
  ARL0 = 360,
  side = "two-sided",
  max.length = 5000,
  nsim.chart = 10000,
  tol.chart = 1e-06,
  plot = TRUE
)
}
\arguments{
\item{Y}{is a vector.}

\item{H}{is the design matrix for shifts.}

\item{X}{is the input matrix}

\item{Y0}{is the initial Y}

\item{q}{is the number of lags.}

\item{A}{is a given variance-covariance matrix in MT and regression for the slab-and-spike coefficients.}

\item{a}{is a given shape of the prior gamma distribution for sigma2.}

\item{b}{is a given scale of the prior gamma distribution for sigma2.}

\item{alpha}{is a given shape of the prior gamma distribution for lambda2.}

\item{beta}{is a given scale of the prior gamma distribution for lambda2.}

\item{theta1}{is a given shape1 of the prior beta distribution for the probability of Tau and Kappa.}

\item{theta2}{is a given shape2 of the prior beta distribution for the probability of Tau and Kappa.}

\item{xi2}{is a given variance of the prior normal distribution for shifts.}

\item{method}{is a choice of methods including MT(McCulloch-Tsay), regression, LASSO, ALASSO(Adaptive LASSO), MonoLASSO(LASSO with Monotonicity constrains), MonoALASSO(Adaptive LASSO with Monotonicity constrains).}

\item{bound0}{is an upper bound of the methods with Monotonicity constrains.}

\item{boundqplus1}{is  a lower bound of the methods with Monotonicity constrains.}

\item{nsim}{is the number of draws from MCMC.}

\item{by}{is the interval of systematic sampling for the draws from MCMC.}

\item{burnin}{is the length of burn-in period.}

\item{tol}{is the tolerance level.}

\item{standardized}{is the flag triggering the standardization for the time series}

\item{logcc}{is the log transformation with continuity correction}

\item{FAP0}{is the given false alarm probability}

\item{estimation.PPP}{is the estimation for Mu0, Phi and sigma2}

\item{nsim.PPP}{is the number of draws for PPP}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- Ph1BayesianLASSO(Y, H = H, q = q, nsim = nsim, burnin = burnin)

}
