% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{PPP}
\alias{PPP}
\title{caculate posterior predictve pvalue}
\usage{
PPP(Y, Phi, Mu, H, TauGamma, sigma2, muq, method = "median", nsim = 3000)
}
\arguments{
\item{Y}{is Y}

\item{Phi}{is the lagged coefficients.}

\item{Mu}{is the individual means}

\item{TauGamma}{is the distributions of shifts.}

\item{sigma2}{is the variance of errors.}

\item{muq}{is the intercept.}

\item{method}{is the method to estimate the parameters.}

\item{nsim}{is the number of simulation for the simulation}
}
\description{
caculate posterior predictve pvalue
}
\examples{
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, q, diag(nrow = q), 0.1, 0.1, 0.1, 0.1, 
1, 1, 0.1, "MonoALASSO", Inf, 0, 1000, 1, 100, 1e-10, H)

PPP(Y0, result$Phi, result$Mu, H, result$Tau * result$Gamma, result$sigma2, result$muq)

}
