% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProdH0.R
\name{ProdH0}
\alias{ProdH0}
\title{obtain the posterior probablity under H0}
\usage{
ProdH0(Tau)
}
\arguments{
\item{Tau}{is the distributions of Tau}
}
\description{
obtain the posterior probablity under H0
}
\examples{
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, q, diag(nrow = q), 0.1, 0.1, 0.1, 0.1, 
1, 1, 0.1, "MonoALASSO", 0, Inf, 1000, 1, 100, 1e-10, H)

ProdH0(result$Tau)
}
