% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinomial.R
\name{ptrinomial}
\alias{ptrinomial}
\title{Get a cdf of trinomial distribution}
\usage{
ptrinomial(x, n, p0)
}
\arguments{
\item{x}{is x}

\item{n}{is trials}

\item{p0}{is ties}
}
\description{
gets the simulated data
}
