% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinomial.R
\name{qtrinomial}
\alias{qtrinomial}
\title{Get a quantile of trinomial distribution}
\usage{
qtrinomial(p, n, p0)
}
\arguments{
\item{p}{is p}

\item{n}{is trials}

\item{p0}{is ties}
}
\description{
gets the simulated data
}
