% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{RMSE}
\alias{RMSE}
\title{obtain the root mean squared error}
\usage{
RMSE(Y, Phi, Mu)
}
\arguments{
\item{Y}{is a vector}

\item{Phi}{is the coefficient}

\item{Mu}{is the mean}
}
\description{
obtain the root mean squared error
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

RMSE(Y, result$Phi, result$Mu)

}
