% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm}
\alias{rmvnorm}
\title{get a sample from multivariate normal distribution}
\usage{
rmvnorm(Mean, Sigma)
}
\arguments{
\item{Mean}{is a mean vector.}

\item{Sigma}{is a variance-covariance matrix.}
}
\description{
get a sample from multivariate normal distribution
}
\examples{
Mean <- rep(0, 5)
Sigma <- diag(nrow = 5)
rmvnorm(Mean, Sigma)
}
