% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rrinvgauss}
\alias{rrinvgauss}
\title{get a sample from inv gaussian distribution}
\usage{
rrinvgauss(n, mu, lambda)
}
\arguments{
\item{n}{is sample size.}

\item{mu}{is a parameter.}

\item{lambda}{is a paremter.}
}
\description{
get a sample from inv gaussian distribution
}
\examples{
rrinvgauss(10, 1, 1)
}
