% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{RSS}
\alias{RSS}
\title{obtain resiudal sum squares}
\usage{
RSS(Y, Phi, Mu)
}
\arguments{
\item{Y}{is a vector}

\item{Phi}{is the coefficient}

\item{Mu}{is the mean}
}
\description{
obtain resiudal sum squares
}
\examples{
nsim <- 100
burnin <- 100
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, H = H, q = q, nsim = nsim, burnin = burnin)

RSS(Y, result$Phi, result$Mu)

}
