% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtrnorm}
\alias{rtrnorm}
\title{Absolute-value-constrained normal distribution}
\usage{
rtrnorm(n, mean, sd, lower, upper)
}
\arguments{
\item{n}{is sample size.}
}
\description{
gets a sample from a normal distribution whose absolute observations are constrained.
}
\examples{
rtwosegnorm(10, 1, 2, 0, 1)
}
