% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtwosegnorm}
\alias{rtwosegnorm}
\title{Absolute-value-constrained normal distribution}
\usage{
rtwosegnorm(n, a, b, mean, sd)
}
\arguments{
\item{n}{is sample size.}

\item{a}{is the lower bound in absolute value.}

\item{b}{is the upper bound in absolute value.}

\item{mean}{is a mean.}

\item{sd}{is a standard deviation.}
}
\description{
gets a sample from a normal distribution whose absolute observations are constrained.
}
\examples{
rtwosegnorm(10, 1, 2, 0, 1)
}
