% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simYXph1}
\alias{simYXph1}
\title{Absolute-value-constrained normal distribution}
\usage{
simYXph1(Y, Phi, Mu, sigma2, theta, eps, leftcensoring, rounding, Z = NULL)
}
\arguments{
\item{n}{is sample size.}
}
\description{
gets a sample from a normal distribution whose absolute observations are constrained.
}
\examples{
rtwosegnorm(10, 1, 2, 0, 1)
}
