% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simYyjph2}
\alias{simYyjph2}
\title{Absolute-value-constrained normal distribution}
\usage{
simYyjph2(h, Yyjph1, Phi, Mu, sigma2)
}
\arguments{
\item{n}{is sample size.}
}
\description{
gets a sample from a normal distribution whose absolute observations are constrained.
}
\examples{
rtwosegnorm(10, 1, 2, 0, 1)
}
