% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{trans}
\alias{trans}
\title{Transform the data}
\usage{
trans(Y, log = TRUE, const = 1, sta = TRUE, meanY = NULL, sdY = NULL)
}
\arguments{
\item{Y}{is the input}

\item{log}{is the flag triggering the log transformation}

\item{const}{is the constant added to the input during the log transformation}

\item{sta}{is the flag triggering the standardization}
}
\description{
gets the transformed input
}
\examples{
alpha <- c(0.03083069, 0.06242601, 0.09120189)
lambda <- 0.239385
pi <- 0.1453097

TT <- 183
w <- 28
Y <- rzinpoisinar3(TT + w, alpha, lambda, pi, ceiling(TT / 2) + w, delta = 1, burnin = 100)
ma <- movaver(Y, w)
ma.tr <- trans(ma, TRUE, 1, TRUE)
}
