% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{yeojohnsontr}
\alias{yeojohnsontr}
\title{Absolute-value-constrained normal distribution}
\usage{
yeojohnsontr(Y, theta, eps)
}
\arguments{
\item{n}{is sample size.}
}
\description{
gets a sample from a normal distribution whose absolute observations are constrained.
}
\examples{
rtwosegnorm(10, 1, 2, 0, 1)
}
