% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring-package.R
\docType{package}
\name{BayesianLASSOMonitoring-package}
\alias{BayesianLASSOMonitoring-package}
\alias{_PACKAGE}
\title{BayesianLASSOMonitoring: What the package does (short line)}
\description{
More about what it does (maybe more than one line)

Package description.
}
\keyword{package}
