% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring-package.R
\docType{package}
\name{BayesianLASSOMonitoring-package}
\alias{BayesianLASSOMonitoring-package}
\alias{_PACKAGE}
\title{BayesianLASSOMonitoring: Random Flexible Level Shift Model based Phase I Control Chart with Bayesian LASSO Estimators}
\description{
This package provides a Phase I Control Chart which is built using Random Flexible Level Shift Model. The model is trained using a Bayesian Technique.

Package description.
}
\keyword{package}
