% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLassoMonitoring-package.R
\docType{package}
\name{BayesianLassoMonitoring-package}
\alias{BayesianLassoMonitoring-package}
\alias{_PACKAGE}
\title{BayesianLassoMonitoring: What the package does (short line)}
\description{
More about what it does (maybe more than one line)

Package description.
}
\keyword{package}
