\name{BayesianLassoMonitoring-package}
\alias{BayesianLassoMonitoring-package}
\alias{BayesianLassoMonitoring}
\docType{package}
\title{
\packageTitle{BayesianLassoMonitoring}
}
\description{
\packageDescription{BayesianLassoMonitoring}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesianLassoMonitoring}
\packageIndices{BayesianLassoMonitoring}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BayesianLassoMonitoring}

Maintainer: \packageMaintainer{BayesianLassoMonitoring}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
