% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHMatGradual}
\alias{getHMatGradual}
\title{get a design matrix for gradual shift}
\usage{
getHMatGradual(T, q, w)
}
\arguments{
\item{T}{is length of a process.}

\item{q}{is the number of lags.}

\item{w}{is the subgroup size.}
}
\description{
get a design matrix for gradual shift
}
\examples{
getHMatGradual(100, 5, 1)
}
