% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHMatIsolated}
\alias{getHMatIsolated}
\title{Design Matrix for Isolated Shift (CM)}
\usage{
getHMatIsolated(T, q)
}
\arguments{
\item{T}{is length of a process.}

\item{q}{is the number of lags.}
}
\description{
gets a design matrix for isolated shift as that in CM
}
\examples{
getHMatIsolated(100, 5)
}
