% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHMatSustained}
\alias{getHMatSustained}
\title{Design Matrix for Sustained Shift (CM)}
\usage{
getHMatSustained(T, q)
}
\arguments{
\item{T}{is length of a process.}

\item{q}{is the number of lags.}
}
\description{
gets a design matrix for sustained shift as that in CM
}
\examples{
getHMatSustained(100, 5)
}
