% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalueKSRFLSM.R
\name{getPvalueKSRFLSM}
\alias{getPvalueKSRFLSM}
\title{get the P value for RFLSM with kernel smoothing}
\usage{
getPvalueKSRFLSM(TauGamma, tail = "2-sided")
}
\arguments{
\item{TauGamma}{is the distributions of shifts.}

\item{tail}{is type of the test.}
}
\description{
get the P value for RFLSM with kernel smoothing
}
\examples{
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, q, diag(nrow = q), 0.1, 0.1, 0.1, 0.1, 1, 1, 0.1, "MonoALASSO", 0, Inf, 1000, 1, 100, 1e-10, H)

getPvalueKSRFLSM(result$Tau * result$Gamma)
}
