% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPvalue.R
\name{getPvalueRFLSM}
\alias{getPvalueRFLSM}
\title{get the P value for RFLSM}
\usage{
getPvalueRFLSM(TauGamma, tail = "2-sided")
}
\arguments{
\item{TauGamma}{is the distributions of shifts.}

\item{tail}{is type of the test.}
}
\description{
get the P value for RFLSM
}
