% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.R
\name{ll}
\alias{ll}
\title{obtain the log likelihood}
\usage{
ll(Y, Phi, Mu, sigma2)
}
\arguments{
\item{Y}{is a vector}

\item{Phi}{is the coefficient}

\item{Mu}{is the mean}

\item{sigma2}{is the variance}
}
\description{
obtain the log likelihood
}
\examples{
T <- 100
q <- 5
H <- getHMatMT(T, q)
Y <- arima.sim(list(ar = 0.5), n = T)

result <- GibbsRFLSM(Y, q, diag(nrow = q), 0.1, 0.1, 0.1, 0.1, 1, 1, 0.1, "MonoALASSO", 0, Inf, 1000, 1, 100, 1e-10, H)

ll(Y, result$Phi, result$Mu, result$sigma2)
}
