% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLASSOMonitoring.R
\name{Ph1MultipleTesting.GammaNormBC}
\alias{Ph1MultipleTesting.GammaNormBC}
\title{Bayesian LASSO Phase I Monitoring 11}
\usage{
Ph1MultipleTesting.GammaNormBC(
  model,
  w = 7,
  FAP0 = 0.05,
  method = "bonferroni",
  side = "right-sided",
  leftcensoring = 1,
  rounding = 0
)
}
\arguments{
\item{model}{is model.}

\item{FAP0}{is}

\item{nsim}{is .}

\item{log}{is model.}

\item{const}{is .}

\item{sta}{is}
}
\description{
gets a posterior sample using Gibbs sampling for Random Flexible Level Shift Model
}
